Add wasm64/memory64 CLI codegen and runtime support
- CLI (js/mod.rs, js/binding.rs): BigInt pointer/length conversions,
  memory64 detection, stack pointer shim, helper functions for
  string/array get with Number() wrapping for memory64 pointers
- CLI (wasm_conventions.rs): Accept I64 globals for stack/TLS base
- CLI (wit/incoming.rs, wit/outgoing.rs): WIT adapter pointer types
- Lib (describe.rs): usize/isize as I64 on 64-bit, pointer descriptors
- Lib (externref.rs): u32 table index types (always i32 per spec)
- Lib (convert/impls.rs, slices.rs, traits.rs): wasm64 conversions
- Lib (closure.rs, convert/closures.rs): BigInt closure data pointers

+++ src/convert/closures.rs
@@ -59,7 +59,7 @@ macro_rules! closures {
             fn into_abi(self) -> WasmSlice {
                 unsafe {
                     let (a, b): (usize, usize) = mem::transmute(self);
-                    WasmSlice { ptr: a as u32, len: b as u32 }
+                    WasmSlice { ptr: a, len: b }
                 }
             }
         }
